Separate component to display deepdao organization
